* extract map element container into a file
